Make blog intro / CTA  visually distinguishable (#748)
* Add class to paragraph containing CTA

Add class to the paragraph containing CTA elements. The class is used to target the container paragraph element to create a box to distinguish CTA section from the content.

* Add css properties for paragraphy containing CTA

Added css properties for paragraph element to create a box around the CTA, centering the text and providing padding.

* Add media query for CTA box

Add media query for CTA box for small screens to prevent individual CTA from breaking. Each CTA appears on an individual line and look clean. CTA dividers are not shown on small screens.